scrabble_replace_simple_django/pyproject.toml
[tool.poetry]
name = "technical-assessment-michael"
version = "0.1.0"
description = "An example technical assessment for Michael Frohlich for Tangent IT Solutions as an intermediate Django Consultant for BRSK"
authors = ["Michael Frohlich <mike.frohl@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
requests = "^2.31.0"
djangorestframework = "^3.15.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
